+++ NEWS.md
@@ -1,4 +1,4 @@
-# development
+# 0.1.3
 
 Breaking changes: